Move DofNumbering from PatchTest to separate file
