+++ src/mechanics/dofs/DofNumbering.h
@@ -1,0 +1,30 @@
+#pragma once
+
+#include "base/Group.h"
+
+#include "mechanics/constraintsPde/Constraints.h"
+#include "mechanics/dofs/DofContainer.h"
+#include "mechanics/nodes/NodeSimple.h"
+
+namespace NuTo
+{
+
+namespace DofNumbering
+{
+
+struct DofInfo
+{
+    DofContainer<int> numIndependentDofs;
+    DofContainer<int> numDependentDofs;
+
+    void Merge(DofType dof, DofInfo other)
+    {
+        numIndependentDofs[dof] = other.numIndependentDofs[dof];
+        numDependentDofs[dof] = other.numDependentDofs[dof];
+    }
+};
+
+DofInfo Build(const Groups::Group<NodeSimple>& dofNodes, DofType dof, const ConstraintPde::Constraints& constraints);
+
+} /* DofNumbering */
+} /* NuTo */
