+++ test/mechanics/cell/PatchTest.cpp
@@ -5,6 +5,8 @@
 
 #include "base/Group.h"
 
+#include "mechanics/dofs/DofNumbering.h"
+
 #include "mechanics/mesh/MeshFem.h"
 #include "mechanics/mesh/MeshFemDofConvert.h"
 
@@ -65,40 +67,6 @@ MeshFem QuadPatchTestMesh()
     return mesh;
 }
 
-struct DofInfo
-{
-    DofContainer<int> numIndependentDofs;
-    DofContainer<int> numDependentDofs;
-};
-
-DofInfo ManualDofNumbering(MeshFem* rMesh, DofType dof)
-{
-    // some manual dof numbering ...
-    Group<NodeSimple> allNodes = rMesh->NodesTotal(dof);
-    Group<NodeSimple> nodesConstrainedInX = rMesh->NodesAtAxis(eDirection::X, dof);
-    Group<NodeSimple> nodesConstrainedInY = Group<NodeSimple>(rMesh->NodeAtCoordinate(Eigen::Vector2d(0, 0), dof));
-
-    Group<NodeSimple> nodesUnconstrainedInX = Difference(allNodes, nodesConstrainedInX);
-    Group<NodeSimple> nodesUnconstrainedInY = Difference(allNodes, nodesConstrainedInY);
-
-    int dofNumber = 0;
-
-    for (auto& node : nodesUnconstrainedInX)
-        node.SetDofNumber(0, dofNumber++);
-    for (auto& node : nodesUnconstrainedInY)
-        node.SetDofNumber(1, dofNumber++);
-
-    for (auto& node : nodesConstrainedInX)
-        node.SetDofNumber(0, dofNumber++);
-    for (auto& node : nodesConstrainedInY)
-        node.SetDofNumber(1, dofNumber++);
-
-    DofInfo dofInfo;
-    dofInfo.numIndependentDofs[dof] = nodesUnconstrainedInX.Size() + nodesUnconstrainedInY.Size();
-    dofInfo.numDependentDofs[dof] = nodesConstrainedInX.Size() + nodesConstrainedInY.Size();
-    return dofInfo;
-}
-
 ConstraintPde::Constraints DefineConstraints(MeshFem* rMesh, DofType dof)
 {
     ConstraintPde::Constraints constraints;
@@ -117,53 +85,6 @@ ConstraintPde::Constraints DefineConstraints(MeshFem* rMesh, DofType dof)
     return constraints;
 }
 
-//! @brief build dof numbering, starting at 0, for all `nodes` regardless of constraints 
-//! @return total number of dofs in `nodes`
-int InitialUnconstrainedNumbering(const Groups::Group<NodeSimple>& nodes)
-{
-    int dofNumber = 0;
-    for (auto& node : nodes)
-        for (int iComponent = 0; iComponent < node.GetNumValues(); ++iComponent)
-            node.SetDofNumber(iComponent, dofNumber++);
-    return dofNumber;
-}
-
-std::vector<bool> FindConstrainedDofs(const ConstraintPde::Constraints& constraints, DofType dof, int numDofs)
-{
-    std::vector<bool> isConstrained(numDofs, false);
-    for (int iEquation = 0; iEquation < constraints.GetNumEquations(dof); ++iEquation)
-    {
-        int dependentDofNumber = constraints.GetEquation(dof, iEquation).GetDependentDofNumber();
-        isConstrained[dependentDofNumber] = true;
-    }
-    return isConstrained;
-}
-
-DofInfo AutomaticDofNumbering(Groups::Group<NodeSimple> dofNodes, DofType dof, const ConstraintPde::Constraints& constraints)
-{
-    int numDependentDofs = constraints.GetNumEquations(dof);
-
-    const int numDofs = InitialUnconstrainedNumbering(dofNodes);
-      
-    std::vector<bool> isConstrained = FindConstrainedDofs(constraints, dof, numDofs);
-   
-    int countIndependentDofs = 0;
-    int countDependentDofs = numDofs - numDependentDofs;
-    for (auto& node : dofNodes)
-        for (int iComponent = 0; iComponent < node.GetNumValues(); ++iComponent)
-        {
-            int dofNumber = node.GetDofNumber(iComponent);
-            if (isConstrained[dofNumber])
-                node.SetDofNumber(iComponent, countDependentDofs++);
-            else
-                node.SetDofNumber(iComponent, countIndependentDofs++);
-        }
-
-    DofInfo dofInfo;
-    dofInfo.numDependentDofs[dof] = numDependentDofs;
-    dofInfo.numIndependentDofs[dof] = numDofs - numDependentDofs;
-    return dofInfo;
-}
 
 BOOST_AUTO_TEST_CASE(PatchTest)
 {
@@ -174,7 +95,7 @@ BOOST_AUTO_TEST_CASE(PatchTest)
     AddDofInterpolation(&mesh, displ, interpolation);
 
     auto constraints = DefineConstraints(&mesh, displ);
-    DofInfo dofInfo = AutomaticDofNumbering(mesh.NodesTotal(displ), displ, constraints);
+    DofNumbering::DofInfo dofInfo = DofNumbering::Build(mesh.NodesTotal(displ), displ, constraints);
 
 
     // ************************************************************************
